2026/01/24 - 011

+++ blog/ninja.py
@@ -1,5 +1,6 @@
 from django.shortcuts import get_object_or_404
 from ninja import PatchDict, Router
+from ninja.errors import HttpError
 
 from blog.models import Article
 from blog.schemas import ArticleIn, ArticleOut
@@ -19,6 +20,9 @@ def get_article(request, article_id: int):
 
 @router.post("/articles", response={201: ArticleOut})
 def create_article(request, payload: ArticleIn):
+    if not request.auth.has_perm("blog.add_article"):
+        raise HttpError(403, "你沒有權限新增文章")
+
     article = Article.objects.create(
         **payload.dict(),
         created_by=request.auth,
@@ -28,6 +32,9 @@ def create_article(request, payload: ArticleIn):
 
 @router.put("/articles/{article_id}", response=ArticleOut)
 def update_article(request, article_id: int, payload: ArticleIn):
+    if not request.auth.has_perm("blog.change_article"):
+        raise HttpError(403, "你沒有權限編輯文章")
+
     article = get_object_or_404(Article, id=article_id)
     for attr, value in payload.dict().items():
         setattr(article, attr, value)
@@ -42,6 +49,9 @@ def partial_update_article(
     article_id: int,
     payload: PatchDict[ArticleIn],
 ):
+    if not request.auth.has_perm("blog.change_article"):
+        raise HttpError(403, "你沒有權限編輯文章")
+
     article = get_object_or_404(Article, id=article_id)
     for attr, value in payload.items():
         setattr(article, attr, value)
@@ -52,6 +62,9 @@ def partial_update_article(
 
 @router.delete("/articles/{article_id}", response={204: None})
 def delete_article(request, article_id: int):
+    if not request.auth.has_perm("blog.delete_article"):
+        raise HttpError(403, "你沒有權限刪除文章")
+
     article = get_object_or_404(Article, id=article_id)
     article.delete()
     return 204, None
